requirements.txt
aiohttp
alembic
bcrypt
beautifulsoup4
broadcaster
colorama
fastapi
httpx
influxdb
influxdb-client
numpy
Pillow
psycopg2
pydantic
pymongo
pytest
pyvirtualdisplay
python-dotenv
requests
SQLAlchemy
torch
uvicorn